chatboatKnBase/requirements.txt
faiss-cpu
sentence-transformers
scikit-learn
matplotlib
beautifulsoup4
requests
schedule
numpy
google-generativeai
python-dotenv